Refactor import paths and update link service type
